Move preferences to separate file
